Move pipeline.ts into src/shared/ so the docker build resolves it
The Deploy workflow has been failing since #57 merged: the client-builder
stage's tsc/vite couldn't resolve `import from '../../../src/pipeline'`
because the Dockerfile only copies src/shared/ into the frontend tree.

The fix could be a one-line `COPY src/pipeline.ts ../src/pipeline.ts`,
but the deeper issue is that pipeline.ts IS shared now — both the
backend (upload, reprocess, routes) and the frontend (previewPipeline)
import it. The "src/shared/" directory was created exactly for this,
with task-engine.ts as its first inhabitant. pipeline.ts belongs there
too.

This commit moves it, updates all callers, and leaves the Dockerfile's
existing `COPY src/shared/` line to pick up both files automatically —
no per-file enumeration that future contributors would have to keep in
sync.

Files:
- src/pipeline.ts        → src/shared/pipeline.ts
- src/pipeline.test.ts   → src/shared/pipeline.test.ts (co-located test)
- src/shared/pipeline.ts: self-import `./shared/task-engine` →
  `./task-engine`; re-export comment updated to match the new home.
- src/upload.ts, src/reprocess.ts: `./pipeline` → `./shared/pipeline`
- src/routes/leagues.ts: `../pipeline` → `../shared/pipeline`
- tests/goal-line.test.ts: `../src/pipeline` → `../src/shared/pipeline`
- frontend/src/lib/previewPipeline.ts:
  `../../../src/pipeline` → `../../../src/shared/pipeline`
- frontend/package.json: add `igc-parser` as a devDependency. This is
  still needed even after the move because the docker client-builder
  has no parent node_modules to walk into — Vite must find igc-parser
  in /build/frontend/node_modules. (Locally it works via parent
  resolution from the repo root.)

Verification:
- npx vitest run: 245/245 ✓
- frontend vitest: 29/29 ✓
- npm run typecheck: clean
- frontend npm run build: green (1574 KB main chunk, same as pre-bug)
