+++ frontend/package.json
@@ -31,6 +31,7 @@
     "@types/react-dom": "^18.2.19",
     "@vitejs/plugin-react": "^4.2.1",
     "@vitest/ui": "^4.1.0",
+    "igc-parser": "^2.0.0",
     "jsdom": "^28.1.0",
     "typescript": "^5.3.3",
     "vite": "^5.1.4",
